Added hack-ish version of magics: %plot, %locals, %globals
%plot [--width width] [--scale height] [--scale SCALE] code

This displays the plots made by Stata with height and width as
specified, scaled by SCALE. This makes it feasible for the user to
request a plot be displayed even if it is embedded in a program or a
fenced block. `code` and also print anything and the plot will still be
displayed (so this overrides the Hydrogen compatibility check).

%globals [-t | --truncate] [regexp?]

Print all globals matching `regexp?` and their contents. If `regexp` is
not specified, all globals are printed. If `--truncate` is specified,
only the first line of each global's contents, as truncated by Stata's
`macro dir`.

%locals [-t | --truncate] [regexp?]

As above, but for locals.